Add files via upload
Created a 2D array for the data. I had to round down the Mass to an integer, because the double and float would error out. The rest of the code outside of the array was to make sure that the array filled properly. Check to see if this will work.